backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
torch==2.3.0
torchvision==0.18.0
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
gdown==5.2.0
python-multipart==0.0.12